% Generated by roxygen2: do not edit by hand
% Please edit documentation in ShinyDrawerCode.Rmd.
\name{data_to_json}
\alias{data_to_json}
\title{Interactive drawing graphic}
\usage{
data_to_json(data)
}
\arguments{
\item{data}{data for graph}
}
\description{
Interactive drawing graphic
}