% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_drawer.R
\name{shiny_drawer}
\alias{shiny_drawer}
\title{Interactive drawing graphic}
\usage{
shiny_drawer(
  data,
  linear = "true",
  draw_start = NULL,
  points_end = NULL,
  x_by = 0.25,
  free_draw = T,
  points = "partial",
  aspect_ratio = 1.5,
  title = "",
  x_range = NULL,
  y_range = NULL,
  x_lab = "",
  y_lab = "",
  drawn_line_color = "steelblue",
  data_tab1_color = "steelblue",
  x_axis_buffer = 0.01,
  y_axis_buffer = 0.05,
  show_finished = T,
  shiny_message_loc = NULL
)
}
\arguments{
\item{data}{data for graph}
}
\description{
Interactive drawing graphic
}
